Interpolate Surface x,y data at 0.3 m spacing if x is too coarse.
Users may supply very coarse x data for surfaces. This is problematic
because taking the numerical derivatives to calculate slope and
curvature of coarse x spacing may result in poorly defined derivatives.
This change resamples the provided x data if it is too coarse. The
original data provided to the Surface object are not preserved in this
case.